Refactor CgTestClassConstructor
