+++ utbot-js/src/main/kotlin/framework/codegen/model/constructor/tree/JsCgVariableConstructor.kt
@@ -9,13 +9,14 @@ import framework.api.js.JsPrimitiveModel
 import org.utbot.framework.codegen.domain.context.CgContext
 import org.utbot.framework.codegen.domain.models.CgLiteral
 import org.utbot.framework.codegen.domain.models.CgValue
+import org.utbot.framework.codegen.tree.Cgcomponents
 import org.utbot.framework.codegen.tree.CgTestClassConstructor
 import org.utbot.framework.codegen.tree.CgVariableConstructor
 import org.utbot.framework.codegen.util.nullLiteral
 
 class JsCgVariableConstructor(ctx: CgContext) : CgVariableConstructor(ctx) {
 
-    private val nameGenerator = CgTestClassConstructor.CgComponents.getNameGeneratorBy(ctx)
+    private val nameGenerator = CgComponents.getNameGeneratorBy(ctx)
 
     override fun getOrCreateVariable(model: UtModel, name: String?): CgValue {
         val baseName = name ?: nameGenerator.nameFrom(model.classId)
