+++ utbot-python/src/main/kotlin/org/utbot/python/framework/codegen/model/constructor/tree/PythonCgTestClassConstructor.kt
@@ -3,10 +3,10 @@ package org.utbot.python.framework.codegen.model.constructor.tree
 import org.utbot.framework.codegen.domain.models.TestClassModel
 import org.utbot.framework.codegen.domain.context.CgContext
 import org.utbot.framework.codegen.domain.models.CgClassFile
-import org.utbot.framework.codegen.tree.CgTestClassConstructor
+import org.utbot.framework.codegen.tree.CgSimpleTestClassConstructor
 import org.utbot.framework.codegen.tree.buildClassFile
 
-internal class PythonCgTestClassConstructor(context: CgContext) : CgTestClassConstructor(context) {
+internal class PythonCgTestClassConstructor(context: CgContext) : CgSimpleTestClassConstructor(context) {
     override fun construct(testClassModel: TestClassModel): CgClassFile {
         return buildClassFile {
             this.declaredClass = withTestClassScope {
